requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
dlib==19.22.1
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
kiwisolver==1.3.2
lpips==0.1.4
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-python-headless==4.5.3.56
Pillow==8.3.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.8.30
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
ninja
# package location
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cu102
torchvision==0.10.1


